Finalizado primeiro retorno ## foi finalizado o metodo BuscarTodosEquipamentosCadastrados na controller. faltando pequenos ajustes que sera integrado nas proximas features. ##

+++ GG_LABOPS_API_CONTROLE/GG_LabOps_Domain/Entities/Laboratory.cs
@@ -5,7 +5,7 @@
         public string? Inventory { get; set; }
         public string? Hostname { get; set; }
         public BrandEquipament BrandEquipament { get; set; }
-        public int BrandEquipamentId { get; set; }
+        public long BrandEquipamentId { get; set; }
         public string SerieNumber { get; set; }
         public string EquipamentType { get; set; }
         public string EquipamentModel { get; set; }
